* enum.c (zip_ary): should use long.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@23496 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ enum.c
@@ -1456,7 +1456,7 @@ zip_ary(VALUE val, NODE *memo, int argc, VALUE *argv)
 {
     volatile VALUE result = memo->u1.value;
     volatile VALUE args = memo->u2.value;
-    int n = memo->u3.cnt++;
+    long n = memo->u3.cnt++;
     volatile VALUE tmp;
     int i;
